% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_checking.R
\name{IMI}
\alias{IMI}
\title{Calculate IMI}
\usage{
IMI(state, k, w = NULL)
}
\arguments{
\item{state}{A topic model state file}

\item{k}{The topic to calculate IMI}

\item{w}{A vector of types to calculate IMI for (faster to calculate)}
}
\description{
Function to calculate instantanueous mutual information for types for a given topic.

See reference for details.
}
\details{

}
\references{
Mimno, D. and Blei, D. Bayesian Checking for Topic Models
}

