% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_summaries.R
\name{type_probability}
\alias{KR1_weight}
\alias{KR2_weight}
\alias{relevance_weight}
\alias{topic_probability}
\alias{type_probability}
\title{Topic model key term summaries}
\usage{
type_probability(state, j, beta = 0)

topic_probability(state, j, beta = 0)

KR1_weight(state, j, beta = 0)

KR2_weight(state, j, beta = 0)

relevance_weight(state, j, beta = 0, lambda = 0.6)
}
\arguments{
\item{state}{A tidy topic model state file}

\item{j}{The number of top words to return}

\item{beta}{Beta hyper parameter. Default is 0 (no smoothing).}

\item{lambda}{Relevance weight. Default is 0.6.}
}
\value{
Returns a data_frame with topic and top terms
}
\description{
The classical way of representing topics by order by the probability for a type within a topic.
}
\details{
To save space the calculations are done using a sparse format,
only returning values for type-topic combination that exist in the model.
This means that unless \code{beta} is set to 0, the returning probabilities
will not sum to 1.

Not all reweighting schemes return a probability (such as KR1, KR2 and relevance)
}
\references{
Topic and keyword re-ranking for LDA-based topic modeling (2009)
LDAvis: A method for visualizing and interpreting topics (2014)
}

